+++ src/app/clusters/zone-management-server/ZoneInformationStorage.h
@@ -1,0 +1,66 @@
+/**
+ *
+ *    Copyright (c) 2025-2026 Project CHIP Authors
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ *
+ */
+
+#pragma once
+
+#include <app-common/zap-generated/cluster-objects.h>
+#include <app/clusters/zone-management-server/TwoDCartesianZoneStorage.h>
+
+namespace chip {
+namespace app {
+namespace Clusters {
+namespace ZoneManagement {
+
+using ZoneInformationStruct    = Structs::ZoneInformationStruct::Type;
+using ZoneTriggerControlStruct = Structs::ZoneTriggerControlStruct::Type;
+
+struct ZoneInformationStorage : ZoneInformationStruct
+{
+    ZoneInformationStorage(){};
+
+    ZoneInformationStorage(const uint16_t & aZoneID, ZoneTypeEnum aZoneType, ZoneSourceEnum aZoneSource,
+                           const Optional<TwoDCartesianZoneStorage> & aTwoDCartZoneStorage)
+    {
+        Set(aZoneID, aZoneType, aZoneSource, aTwoDCartZoneStorage);
+    }
+
+    ZoneInformationStorage(const ZoneInformationStorage & aZoneInfoStorage) { *this = aZoneInfoStorage; }
+
+    ZoneInformationStorage & operator=(const ZoneInformationStorage & aZoneInfoStorage)
+    {
+        Set(aZoneInfoStorage.zoneID, aZoneInfoStorage.zoneType, aZoneInfoStorage.zoneSource, aZoneInfoStorage.twoDCartZoneStorage);
+        return *this;
+    }
+
+    void Set(const uint16_t & aZoneID, ZoneTypeEnum aZoneType, ZoneSourceEnum aZoneSource,
+             const Optional<TwoDCartesianZoneStorage> & aTwoDCartZoneStorage)
+    {
+        zoneID              = aZoneID;
+        zoneType            = aZoneType;
+        zoneSource          = aZoneSource;
+        twoDCartZoneStorage = aTwoDCartZoneStorage;
+        twoDCartesianZone   = twoDCartZoneStorage;
+    }
+
+    Optional<TwoDCartesianZoneStorage> twoDCartZoneStorage;
+};
+
+} // namespace ZoneManagement
+} // namespace Clusters
+} // namespace app
+} // namespace chip
