+++ src/app/clusters/zone-management-server/ZoneManagementCluster.h
@@ -23,7 +23,8 @@
 #include <app/server-cluster/DefaultServerCluster.h>
 #include <lib/core/CHIPError.h>
 #include <protocols/interaction_model/StatusCode.h>
-
+#include <app/clusters/zone-management-server/TwoDCartesianZoneStorage.h>
+#include <app/clusters/zone-management-server/Delegate.h>
 #include <optional>
 #include <string>
 #include <vector>
@@ -32,115 +33,6 @@ namespace chip {
 namespace app {
 namespace Clusters {
 namespace ZoneManagement {
-
-using TwoDCartesianZoneDecodableStruct = Structs::TwoDCartesianZoneStruct::DecodableType;
-using TwoDCartesianZoneStruct          = Structs::TwoDCartesianZoneStruct::Type;
-using TwoDCartesianVertexStruct        = Structs::TwoDCartesianVertexStruct::Type;
-using ZoneInformationStruct            = Structs::ZoneInformationStruct::Type;
-using ZoneTriggerControlStruct         = Structs::ZoneTriggerControlStruct::Type;
-
-struct TwoDCartesianZoneStorage : TwoDCartesianZoneStruct
-{
-    TwoDCartesianZoneStorage() = default;
-
-    TwoDCartesianZoneStorage(const CharSpan & aName, ZoneUseEnum aUse, const std::vector<TwoDCartesianVertexStruct> & aVertices,
-                             Optional<CharSpan> aColor)
-    {
-        Set(aName, aUse, aVertices, aColor);
-    }
-
-    TwoDCartesianZoneStorage(const TwoDCartesianZoneStorage & aTwoDCartZone) { *this = aTwoDCartZone; }
-
-    TwoDCartesianZoneStorage & operator=(const TwoDCartesianZoneStorage & aTwoDCartZone)
-    {
-        Set(aTwoDCartZone.name, aTwoDCartZone.use, aTwoDCartZone.verticesVector, aTwoDCartZone.color);
-        return *this;
-    }
-
-    void Set(const CharSpan & aName, ZoneUseEnum aUse, const std::vector<TwoDCartesianVertexStruct> & aVertices,
-             Optional<CharSpan> aColor)
-    {
-        nameString     = std::string(aName.begin(), aName.end());
-        name           = CharSpan(nameString.c_str(), nameString.size());
-        use            = aUse;
-        verticesVector = aVertices;
-        vertices       = DataModel::List<TwoDCartesianVertexStruct>(verticesVector.data(), verticesVector.size());
-        if (aColor.HasValue())
-        {
-            colorString = std::string(aColor.Value().begin(), aColor.Value().end());
-            color       = MakeOptional(CharSpan(colorString.c_str(), colorString.size()));
-        }
-        else
-        {
-            colorString.clear();
-            color = NullOptional;
-        }
-    }
-
-    std::string nameString;
-    std::string colorString;
-    std::vector<TwoDCartesianVertexStruct> verticesVector;
-};
-
-struct ZoneInformationStorage : ZoneInformationStruct
-{
-    ZoneInformationStorage() = default;
-
-    ZoneInformationStorage(const uint16_t & aZoneID, ZoneTypeEnum aZoneType, ZoneSourceEnum aZoneSource,
-                           const Optional<TwoDCartesianZoneStorage> & aTwoDCartZoneStorage)
-    {
-        Set(aZoneID, aZoneType, aZoneSource, aTwoDCartZoneStorage);
-    }
-
-    ZoneInformationStorage(const ZoneInformationStorage & aZoneInfoStorage) { *this = aZoneInfoStorage; }
-
-    ZoneInformationStorage & operator=(const ZoneInformationStorage & aZoneInfoStorage)
-    {
-        Set(aZoneInfoStorage.zoneID, aZoneInfoStorage.zoneType, aZoneInfoStorage.zoneSource, aZoneInfoStorage.twoDCartZoneStorage);
-        return *this;
-    }
-
-    void Set(const uint16_t & aZoneID, ZoneTypeEnum aZoneType, ZoneSourceEnum aZoneSource,
-             const Optional<TwoDCartesianZoneStorage> & aTwoDCartZoneStorage)
-    {
-        zoneID              = aZoneID;
-        zoneType            = aZoneType;
-        zoneSource          = aZoneSource;
-        twoDCartZoneStorage = aTwoDCartZoneStorage;
-        twoDCartesianZone   = twoDCartZoneStorage;
-    }
-
-    Optional<TwoDCartesianZoneStorage> twoDCartZoneStorage;
-};
-
-class ZoneManagementCluster;
-
-class Delegate
-{
-public:
-    Delegate()          = default;
-    virtual ~Delegate() = default;
-
-    virtual Protocols::InteractionModel::Status CreateTwoDCartesianZone(const TwoDCartesianZoneStorage & zone,
-                                                                        uint16_t & outZoneID)                                   = 0;
-    virtual Protocols::InteractionModel::Status UpdateTwoDCartesianZone(uint16_t zoneID, const TwoDCartesianZoneStorage & zone) = 0;
-    virtual Protocols::InteractionModel::Status RemoveZone(uint16_t zoneID)                                                     = 0;
-    virtual Protocols::InteractionModel::Status CreateTrigger(const ZoneTriggerControlStruct & zoneTrigger)                     = 0;
-    virtual Protocols::InteractionModel::Status UpdateTrigger(const ZoneTriggerControlStruct & zoneTrigger)                     = 0;
-    virtual Protocols::InteractionModel::Status RemoveTrigger(uint16_t zoneID)                                                  = 0;
-    virtual void OnAttributeChanged(AttributeId attributeId)                                                                    = 0;
-    virtual CHIP_ERROR PersistentAttributesLoadedCallback()                                                                     = 0;
-    virtual CHIP_ERROR LoadZones(std::vector<ZoneInformationStorage> & aZones)                                                  = 0;
-    virtual CHIP_ERROR LoadTriggers(std::vector<ZoneTriggerControlStruct> & aTriggers)                                          = 0;
-
-protected:
-    friend class ZoneManagementCluster;
-    // This is named mZoneManagementServer instead of mZoneManagementCluster to preserve backwards compatibility with legacy usage.
-    ZoneManagementCluster * mZoneManagementServer = nullptr;
-
-    void SetZoneManagementCluster(ZoneManagementCluster * zoneManagementCluster) { mZoneManagementServer = zoneManagementCluster; }
-    ZoneManagementCluster * GetZoneManagementCluster() const { return mZoneManagementServer; }
-};
 
 class ZoneManagementCluster : public DefaultServerCluster
 {
